+++ crates/ruma-common/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+Bug fixes:
+
+- Add `StatusProfileField::new()` and `CallProfileField::new()` constructors for MSC4426
+
 Improvements:
 
 - Add `From` conversions between `MatrixToUri` and `MatrixUri`.
@@ -9,7 +13,6 @@ Improvements:
 - Add `M_CONCURRENT_WRITE` error code, used by [MSC4438].
 - Add `canonical_json::RedactingSerializer` to serialize a `CanonicalJsonObject`
   while redacting it on the fly.
-- Add `StatusProfileField::new()` and `CallProfileField::new()` constructors for MSC4426
 
 [MSC4438]: https://github.com/matrix-org/matrix-spec-proposals/pull/4438